requirements.txt
fastapi
uvicorn
gunicorn
jinja2
python-multipart
azure-monitor-opentelemetry==1.1.0